Removed legacy viewer and refactored content to @itwinjs-sandbox
