+++ .vscode/settings.json
@@ -13,10 +13,12 @@
     "Metrostation",
     "Oidc",
     "Pictometry",
+    "Walkthrough",
     "bentleyjs",
     "exton",
     "fontawesome",
     "frontstage",
+    "hypermodeling",
     "imodel",
     "imodelhub",
     "imodeljs",
@@ -29,9 +31,12 @@
     "orbitgt",
     "rbac",
     "ruleset",
+    "rulesets",
+    "signin",
     "tileset",
     "typedefs",
     "typemoq",
+    "uicore",
     "webfont",
     "webgl"
   ],
