+++ src/frontend-samples/clash-review-sample/sampleSpec.ts
@@ -13,7 +13,6 @@ export function getClashReviewSpec(): SampleSpec {
     image: "clashreview.png",
     description: "Uses the #validation REST api to get and visualize #clash results.",
     readme: async () => import("!!raw-loader!./readme.md"),
-    iTwinViewerReady: true,
     files: () => [
       import("!editor-file-loader!./ClashReviewApi"),
       import("!editor-file-loader!./ClashReviewApp?entry=true"),
