+++ src/frontend-samples/classifier-sample/sampleSpec.ts
@@ -11,7 +11,6 @@ export function getClassifierSpec(): SampleSpec {
     name: "classifier-sample",
     label: "Classifiers",
     image: "classifier-thumbnail.png",
-    iTwinViewerReady: true,
     readme: async () => import("!!raw-loader!./readme.md"),
     files: () => [
       import("!editor-file-loader!./ClassifierApi"),
