+++ src/frontend-samples/cross-probing-sample/sampleSpec.ts
@@ -12,7 +12,6 @@ export function getCrossProbingSpec(): SampleSpec {
     label: "Cross-Probing",
     image: "cross-probing-thumbnail.png",
     iModelList: [SampleIModels.BayTown],
-    iTwinViewerReady: true,
     readme: async () => import("!!raw-loader!./readme.md"),
     files: () => [
       import("!editor-file-loader!./CrossProbingApi"),
