+++ src/frontend-samples/geometry-samples/advanced-3d-sample/sampleSpec.ts
@@ -12,7 +12,6 @@ export function getAdvanced3dSpec(): SampleSpec {
     image: "advanced-3d-thumbnail.png",
     description: "#Geometry #sample showing how to generate several #3d pieces of geometry including #mitered #pipes, #rotational, #linear, and #ruled, #sweeps.",
     iModelList: [],
-    iTwinViewerReady: true,
     readme: async () => import("!!raw-loader!./readme.md"),
     files: () => [
       import("!editor-file-loader!./Advanced3dApi"),
