+++ src/frontend-samples/geometry-samples/curve-fraction/sampleSpec.ts
@@ -12,7 +12,6 @@ export function getCurveFractionSpec(): SampleSpec {
     image: "curve-fraction-thumbnail.png",
     description: "#Geometry #sample showing how to find a #point along a set #fraction of a #curve using the #fractionToPointAndDerivative method",
     iModelList: [],
-    iTwinViewerReady: true,
     readme: async () => import("!!raw-loader!./readme.md"),
     files: () => [
       import("!editor-file-loader!./CurveFractionApi"),
