+++ src/frontend-samples/iot-alert-sample/sampleSpec.ts
@@ -12,7 +12,6 @@ export function getIotAlertSpec(): SampleSpec {
     label: "IoT Alerts",
     image: "IoT-Alert-thumbnail.png",
     description: "How to simulate out-of-bound condition from a #sensor and trigger #IotAlert, #blinking effect",
-    iTwinViewerReady: true,
     readme: async () => import("!!raw-loader!./readme.md"),
     files: () => [
       import("!editor-file-loader!./IotAlertApi"),
