+++ src/frontend-samples/volume-query-sample/sampleSpec.ts
@@ -12,7 +12,6 @@ export function getVolumeQuerySpec(): SampleSpec {
     label: "Volume Query",
     image: "volume-query-thumbnail.png",
     description: "Query #SpatialElements using #SpatialQueries. Elements are classified using #getGeometryContainment. To color and represent elements #EmphasizeElements and #PresentationLabelsProvider are being used.",
-    iTwinViewerReady: true,
     readme: async () => import("!!raw-loader!./readme.md"),
     files: () => [
       import("!editor-file-loader!./VolumeQueryApi"),
