+++ src/frontend-samples/zoom-to-elements-sample/sampleSpec.ts
@@ -12,7 +12,6 @@ export function getZoomToElementsSpec(): SampleSpec {
     label: "Zoom to Elements",
     image: "zoom-to-elements-thumbnail.png",
     description: "Uses #zoomToElement, #ZoomToOptions, and #ViewChangeOptions to center the view on element(s) of interest.",
-    iTwinViewerReady: true,
     readme: async () => import("!!raw-loader!./readme.md"),
     files: () => [
       import("!editor-file-loader!./ZoomToElementsApi"),
